feat(billing): aligne old billing
