+++ packages/manager/modules/billing-components/src/components/terminate-vrack/component.js
@@ -5,7 +5,6 @@ export default {
   bindings: {
     goBack: '<',
     service: '<',
-    serviceType: '<',
     isEmpty: '<',
   },
   controller,
