+++ packages/manager/modules/billing-components/src/components/terminate-vrack/template.html
@@ -26,7 +26,7 @@
 >
     <oui-message data-type="warning" class="mb-2">
         <span
-            data-translate="{{:: 'autorenew_agora_terminate_service_warning_' + $ctrl.serviceType}}"
+            data-translate="{{:: 'autorenew_agora_terminate_service_warning_VRACK'}}"
         ></span>
     </oui-message>
     <oui-field
