--- packages/manager/modules/billing/src/autoRenew/actions/terminate-vrack/component.js
@@ -1,14 +1,0 @@
-import template from './template.html';
-import controller from './controller';
-
-export default {
-  bindings: {
-    goBack: '<',
-    service: '<',
-    serviceType: '<',
-    isEmpty: '<',
-  },
-  controller,
-  template,
-  name: 'billingAutorenewTerminateVrack',
-};
