--- packages/manager/modules/billing/src/autoRenew/actions/terminate-vrack/constants.js
@@ -1,5 +1,0 @@
-export const TERMINATE_PATTERN = /^TERMINATE$/;
-
-export default {
-  TERMINATE_PATTERN,
-};
