+++ packages/manager/modules/billing/src/autoRenew/actions/terminate-vrack/index.js
@@ -5,9 +5,6 @@ import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
 import '@ovh-ux/ui-kit';
 import uiRouter from '@uirouter/angularjs';
 
-import component from './component';
-import routing from './routing';
-
 const moduleName = 'ovhManagerBillingAutorenewTerminateVrack';
 
 angular
@@ -18,8 +15,42 @@ angular
     'oui',
     uiRouter,
   ])
-  .config(routing)
-  .component(component.name, component)
+  .config(
+    /* @ngInject */ ($stateProvider) => {
+      $stateProvider.state('billing.autorenew.terminateVrack', {
+        url: '/terminate-vrack?service&serviceType',
+        views: {
+          modal: {
+            component: 'billingTerminateVrack',
+          },
+        },
+        layout: 'modal',
+        resolve: {
+          goBack: /* @ngInject */ (goToAutorenew) => goToAutorenew,
+          service: /* @ngInject */ ($transition$) =>
+            $transition$.params().service,
+          isEmpty: /* @ngInject */ (OvhApiVrack, service) =>
+            OvhApiVrack.Aapi()
+              .services({ serviceName: service })
+              .$promise.then((allServicesParam) => {
+                const services = Object.entries(allServicesParam).filter(
+                  ([, value]) => {
+                    return Array.isArray(value) && value.length;
+                  },
+                );
+                return !services.length;
+              })
+              .catch(() => {
+                return false;
+              }),
+          breadcrumb: () => null,
+        },
+        atInternet: {
+          ignore: true,
+        },
+      });
+    },
+  )
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
