--- packages/manager/modules/billing/src/autoRenew/actions/terminate-vrack/translations/Messages_pt_PT.json
@@ -1,9 +1,0 @@
-{
-  "autorenew_agora_terminate_vrack_blocked_title": "Eliminar o meu vRack",
-  "autorenew_agora_terminate_vrack_blocked_description": "Para eliminar o seu vRack, deve desligar previamente todos os serviços que estão associados ao mesmo.",
-  "autorenew_agora_terminate_vrack_blocked_close": "Fechar",
-  "autorenew_agora_terminate_service_VRACK": "Eliminar o vRack",
-  "autorenew_agora_terminate_service_warning_VRACK": "Confirme a eliminação do vRack",
-  "autorenew_agora_terminate_service_success_VRACK": "O seu pedido de eliminação do seu vRack foi registado. Um e-mail com o procedimento foi-lhe enviado.",
-  "autorenew_agora_terminate_service_error_VRACK": "Ocorreu um erro aquando do pedido de eliminação do seu vRack. {{error}}"
-}
