+++ packages/manager/modules/new-billing/src/autoRenew/actions/terminate-vrack/index.js
@@ -4,7 +4,6 @@ import ngAtInternet from '@ovh-ux/ng-at-internet';
 import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
 import '@ovh-ux/ui-kit';
 import uiRouter from '@uirouter/angularjs';
-import routing from './routing';
 
 const moduleName = 'ovhManagerBillingAutorenewTerminateVrack';
 
@@ -16,7 +15,42 @@ angular
     'oui',
     uiRouter,
   ])
-  .config(routing)
+  .config(
+    /* @ngInject */ ($stateProvider) => {
+      $stateProvider.state('billing.autorenew.terminateVrack', {
+        url: '/terminate-vrack?service&serviceType',
+        views: {
+          modal: {
+            component: 'billingTerminateVrack',
+          },
+        },
+        layout: 'modal',
+        resolve: {
+          goBack: /* @ngInject */ (goToAutorenew) => goToAutorenew,
+          service: /* @ngInject */ ($transition$) =>
+            $transition$.params().service,
+          isEmpty: /* @ngInject */ (OvhApiVrack, service) =>
+            OvhApiVrack.Aapi()
+              .services({ serviceName: service })
+              .$promise.then((allServicesParam) => {
+                const services = Object.entries(allServicesParam).filter(
+                  ([, value]) => {
+                    return Array.isArray(value) && value.length;
+                  },
+                );
+                return !services.length;
+              })
+              .catch(() => {
+                return false;
+              }),
+          breadcrumb: () => null,
+        },
+        atInternet: {
+          ignore: true,
+        },
+      });
+    },
+  )
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
